Spin box: fix the width of arrow buttons
This patch reworks the computation of the spin box' arrow buttons
width. It removes rendering of the "small" variant (two smaller
buttons on top of each other), keeping only the "large" variant
(two larger buttons side-by-side).

The width of the buttons is estimated from the height of the
unexpanded spin box.

The estimated width of the buttons is added to the content width
in the Adwaita::sizeFromContents() to ensure that the minimum
size hint of the spin box always accounts for the buttons size, and
prevent them from covering the contents.

An attempt to address issues #37 and #41.