Adwaita: implemented animated "busy" indicator for QProgressBar
According to QProgressBar docs, "If minimum and maximum both are
set to 0, the bar shows a busy indicator instead of a percentage
of steps".

We use QProgressStyleAnimation to animate the busy indicator,
which makes it move a bit slower than its Gtk counterpart. Still,
better than not indicating the activity at all...